requirements/base.txt
Django==4.1.3
Jinja2==3.0.3
Pillow==9.3.0
djangorestframework==3.13.1
redis==4.3.1
django-redis==5.2.0
celery==5.2.7
environs==9.3.0
psycopg2==2.9.3
django-timestampable==1.0.3
dj_database_url==1.0.0
django-mptt==0.14.0
django-extensions==3.2.1
django-celery-beat==2.4.0
django-celery-results==2.4.0
requests==2.28.2
django-debug-toolbar==3.8.1
